@@ -16,7 +16,7 @@
     <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
     <RepositoryType></RepositoryType>
     <RepositoryUrl></RepositoryUrl>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <LangVersion>7.3</LangVersion>
     <TargetLatestRuntimePatch>true</TargetLatestRuntimePatch>
     <Version>2.0.5</Version>
@@ -24,8 +24,9 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <!--<PackageReference Include="Microsoft.AspNetCore.App"  />-->
-    <PackageReference Include="Microsoft.AspNetCore.App" Version="2.1.0" />
+
+    <PackageReference Include="Microsoft.AspNetCore.App" Version="2.2.7" />
+
     <PackageReference Include="Newtonsoft.Json" Version="12.0.2" />
   </ItemGroup>
 </Project>
